Lisää taas...

+++ Osa4/blogilista/controllers/blogs.js
@@ -1,10 +1,19 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
 // GET / - Get all blogs
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -12,7 +21,12 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const user = await User.findById(body.userId)
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   if (body.title === undefined || body.url === undefined) {
     response.status(400).end()
@@ -24,20 +38,11 @@ blogsRouter.post('/', async (request, response) => {
     author: body.author,
     url: body.url,
     likes: body.likes ? body.likes : 0,
-    user: {
-      username: user.username,
-      name: user.name,
-      id: user.id
-    }
+    user: user._id
   })
 
   const savedBlog = await blog.save()
-  user.blogs = user.blogs.concat({
-    url: blog.url,
-    title: blog.title,
-    author: blog.author,
-    id: savedBlog.id
-  })
+  user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
   response.status(201).json(savedBlog)
@@ -45,8 +50,31 @@ blogsRouter.post('/', async (request, response) => {
 
 // DELETE /:id - Remove a blog post by id
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const body = request.body
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  if (user === null || user === undefined) {
+    return response.status(404).json({ error: 'user not found' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  console.log(blog.user)
+  console.log(user)
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+
+  return response.status(401).json({ error: 'user not same as blog poster' })
 })
 
 // PUT /:id - Update a blog by id
